update brick template
