+++ bricks/flutter_starter/__brick__/{{name}}/lib/api_sdk/dio/dio_helpers/dio_client.dart
@@ -1,8 +1,9 @@
-import 'package:{{name}}/api_sdk/dio/models/user_model.dart';
-import 'package:{{name}}/config/config.dart';
-import 'package:{{name}}/{{name}}.dart';
-import 'package:retrofit/retrofit.dart';
 import 'package:dio/dio.dart';
+import 'package:retrofit/retrofit.dart';
+
+import '../../../config/config.dart';
+import 'dio_apis.dart';
+
 part 'dio_client.g.dart';
 
 @RestApi(baseUrl: baseUrl)
@@ -10,8 +11,11 @@ abstract class DioClient {
   factory DioClient(Dio dio, {String baseUrl}) = _DioClient;
 
   @GET(DioApis.users)
-  Future<DioResponseData> getUsers(@Path("id") int id);
+  Future<dynamic> getUsers(@Path("id") int id);
 
   @POST(DioApis.login)
-  Future<dynamic> createPost(@Body() dynamic body);
+  Future<dynamic> loginUser(@Body() dynamic body);
+
+  @POST(DioApis.register)
+  Future<dynamic> registerUser(@Body() dynamic body);
 }
