+++ bricks/flutter_starter/__brick__/{{name}}/lib/api_sdk/dio/dio_helpers/dio_interceptor.dart
@@ -1,5 +1,6 @@
 import 'package:dio/dio.dart';
-import 'package:{{name}}/{{name}}.dart';
+
+import 'dio_exceptions.dart';
 
 class DioInterceptor extends Interceptor {
   @override
