+++ bricks/flutter_starter/__brick__/{{name}}/lib/api_sdk/dio/models/user_model.dart
@@ -1,4 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
+
 part 'user_model.g.dart';
 
 @JsonSerializable()
