+++ bricks/flutter_starter/__brick__/{{name}}/lib/api_sdk/http/http_helpers/http_client.dart
@@ -1,14 +1,13 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:{{name}}/config/config.dart';
 import 'package:http/http.dart' as http;
-import 'package:http/http.dart';
 
+import '../../../config/config.dart';
 import 'http_exception.dart';
 
 class HttpClient {
-  final Client client;
+  final http.Client client;
   HttpClient({required this.client});
   Future<dynamic> get(
     String url, {
