+++ bricks/flutter_starter/__brick__/{{name}}/lib/api_sdk/http/http_service.dart
@@ -1,7 +1,8 @@
-import 'package:{{name}}/config/config.dart';
-import 'package:{{name}}/{{name}}.dart';
 import 'package:http/http.dart';
 import 'package:http_interceptor/http/intercepted_client.dart';
+
+import 'http_helpers/http_client.dart';
+import 'http_helpers/http_interceptor.dart';
 
 class HttpApi {
   HttpApi._internal();
