+++ bricks/flutter_starter/__brick__/{{name}}/lib/app.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:{{name}}/{{name}}.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
+
+import 'common_export.dart';
 
 class App extends StatelessWidget {
   const App({Key? key}) : super(key: key);
