+++ bricks/flutter_starter/__brick__/{{name}}/lib/common_export.dart
@@ -1,0 +1,27 @@
+export 'api_sdk/{{api}}_api_sdk.dart';
+export 'app.dart';
+export 'routes/routes.dart';
+export 'screens/home/home.dart';
+export 'screens/login/login_screen.dart';
+export 'screens/login/widgets/login_form.dart';
+export 'screens/signup/signup_screen.dart';
+export 'screens/signup/widgets/signup_form.dart';
+export 'screens/splash/splash_screen.dart';
+export 'shared/bloc/authentication/authentication_bloc.dart';
+export 'shared/bloc/bloc_controller.dart';
+export 'shared/bloc/update_theme/update_theme_bloc.dart';
+export 'shared/models/current_user_data.dart';
+export 'shared/models/current_user_data.dart';
+export 'shared/models/firebase_user.dart';
+export 'shared/models/firebase_user.dart';
+export 'shared/models/token.dart';
+export 'shared/models/user.dart';
+export 'shared/models/user_data.dart';
+export 'shared/repository/authentication_repository.dart';
+export 'shared/shared.dart';
+export 'themes/color_constants.dart';
+export 'themes/theme_data.dart';
+export 'utils/image_constants.dart';
+export 'utils/responsive.dart';
+export 'utils/size_utils.dart';
+export 'widgets/cache_image_widget.dart';
