+++ bricks/flutter_starter/__brick__/{{name}}/lib/main.dart
@@ -1,6 +1,7 @@
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:{{name}}/{{name}}.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+import 'common_export.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
