+++ bricks/flutter_starter/__brick__/{{name}}/lib/routes/routes.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:{{name}}/{{name}}.dart';
 import 'package:go_router/go_router.dart';
+
+import '../common_export.dart';
 
 final GoRouter router = GoRouter(
   initialLocation: '/',
@@ -32,10 +33,8 @@ final GoRouter router = GoRouter(
   ],
 );
 
-
-
 /// routes for the app  Without go_router
-// import 'package:{{name}}/{{name}}.dart';
+// import 'package:flutter_starter/flutter_starter.dart';
 
 // import 'package:flutter/material.dart';
