+++ bricks/flutter_starter/__brick__/{{name}}/lib/screens/home/home.dart
@@ -1,10 +1,11 @@
-// home screen contents
-import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
+
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:{{name}}/{{name}}.dart';
 import 'package:go_router/go_router.dart';
+
+import '../../common_export.dart';
 
 class HomeScreen extends StatefulWidget {
   const HomeScreen({Key? key}) : super(key: key);
