+++ bricks/flutter_starter/__brick__/{{name}}/lib/screens/login/widgets/login_form.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:{{name}}/{{name}}.dart';
+
+import '../../../common_export.dart';
 
 class LoginForm extends StatefulWidget {
   final AuthenticationBloc authenticationBloc;
