+++ bricks/flutter_starter/__brick__/{{name}}/lib/screens/signup/widgets/signup_form.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:{{name}}/{{name}}.dart';
+
+import '../../../common_export.dart';
 
 class SignUpForm extends StatefulWidget {
   final AuthenticationBloc authenticationBloc;
