+++ bricks/flutter_starter/__brick__/{{name}}/lib/screens/splash/splash_screen.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:{{name}}/{{name}}.dart';
 import 'package:go_router/go_router.dart';
+
+import '../../common_export.dart';
 
 class SplashScreen extends StatefulWidget {
   const SplashScreen({Key? key}) : super(key: key);
