+++ bricks/flutter_starter/__brick__/{{name}}/lib/shared/bloc/authentication/authentication_bloc.dart
@@ -1,6 +1,7 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:{{name}}/{{name}}.dart';
+
+import '../../../common_export.dart';
 
 part 'authentication_event.dart';
 part 'authentication_state.dart';
