+++ bricks/flutter_starter/__brick__/{{name}}/lib/shared/bloc/update_theme/update_theme_bloc.dart
@@ -1,5 +1,8 @@
-import 'package:bloc/bloc.dart';
-import 'package:{{name}}/{{name}}.dart';
+import 'package:equatable/equatable.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+part 'update_theme_event.dart';
+part 'update_theme_state.dart';
 
 class UpdateThemeBloc extends Bloc<UpdateThemeEvent, UpdateThemeState> {
   UpdateThemeBloc() : super(const SetTheme(appTheme: AppTheme.light)) {
