+++ bricks/flutter_starter/__brick__/{{name}}/lib/shared/bloc/update_theme/update_theme_event.dart
@@ -1,4 +1,4 @@
-import 'package:equatable/equatable.dart';
+part of 'update_theme_bloc.dart';
 
 enum AppTheme { light, dark }
