+++ bricks/flutter_starter/__brick__/{{name}}/lib/shared/bloc/update_theme/update_theme_state.dart
@@ -1,6 +1,4 @@
-import 'package:equatable/equatable.dart';
-
-import '../../../{{name}}.dart';
+part of 'update_theme_bloc.dart';
 
 abstract class UpdateThemeState extends Equatable {
   const UpdateThemeState();
