+++ bricks/flutter_starter/__brick__/{{name}}/lib/shared/repository/authentication_repository.dart
@@ -1,10 +1,9 @@
-import 'package:{{name}}/{{name}}.dart';
+import '../../common_export.dart';
 
 class AuthenticationRepository {
   Future<dynamic> signUpWithEmailAndPassword(
       String email, String password) async {
-    await Future.delayed(
-        const Duration(seconds: 1)); // simulate a network delay
+    await Future.delayed(const Duration(seconds: 1));
     final response = await ApiSdk.signUpWithEmailAndPassword(
         {'email': email, 'password': password});
 
@@ -13,8 +12,7 @@ class AuthenticationRepository {
 
   Future<dynamic> loginWithEmailAndPassword(
       String email, String password) async {
-    await Future.delayed(
-        const Duration(seconds: 1)); // simulate a network delay
+    await Future.delayed(const Duration(seconds: 1));
     final response = await ApiSdk.loginWithEmailAndPassword(
         {'email': email, 'password': password});
