--- bricks/flutter_starter/__brick__/{{name}}/lib/shared/repository/dio_authentication_repository.dart
@@ -1,10 +1,0 @@
-import 'package:{{name}}/api_sdk/dio/dio_service.dart';
-import 'package:{{name}}/api_sdk/dio/models/user_model.dart';
-
-class DioAuthenticationRepository {
-  Future<DioResponseData> getUserData(int id) async {
-    final response = await DioService.getUserData(id);
-
-    return response;
-  }
-}
