--- bricks/flutter_starter/__brick__/{{name}}/lib/{{name}}.dart
@@ -1,37 +1,0 @@
-export 'api_sdk/http/http_service.dart';
-export 'api_sdk/http/http_helpers/http_client.dart';
-export 'api_sdk/http/http_helpers/http_interceptor.dart';
-export 'shared/models/firebase_user.dart';
-export 'shared/bloc/bloc_controller.dart';
-export 'shared/models/firebase_user.dart';
-export 'shared/repository/http_authentication_repository.dart';
-export 'api_sdk/api_sdk.dart';
-export 'shared/models/token.dart';
-export 'shared/models/user.dart';
-export 'shared/models/user_data.dart';
-export 'shared/models/current_user_data.dart';
-export 'shared/shared.dart';
-export 'shared/bloc/authentication/authentication_bloc.dart';
-export 'shared/models/current_user_data.dart';
-export 'app.dart';
-export 'themes/color_constants.dart';
-export 'utils/image_constants.dart';
-export 'themes/theme_data.dart';
-export 'routes/routes.dart';
-export 'screens/login/widgets/login_form.dart';
-export 'screens/signup/widgets/signup_form.dart';
-export '/widgets/cache_image_widget.dart';
-export 'screens/splash/splash_screen.dart';
-export 'screens/home/home.dart';
-export 'api_sdk/dio/dio_helpers/dio_apis.dart';
-export 'api_sdk/dio/dio_helpers/dio_client.dart';
-export 'shared/repository/dio_authentication_repository.dart';
-export 'api_sdk/dio/dio_helpers/dio_exceptions.dart';
-export 'api_sdk/dio/dio_helpers/dio_interceptor.dart';
-export 'package:{{name}}/shared/bloc/update_theme/update_theme_event.dart';
-export 'package:{{name}}/shared/bloc/update_theme/update_theme_state.dart';
-export 'package:{{name}}/shared/bloc/update_theme/update_theme_bloc.dart';
-export 'utils/responsive.dart';
-export 'utils/size_utils.dart';
-export 'screens/login/login_screen.dart';
-export 'screens/signup/signup_screen.dart';
