+++ bricks/flutter_starter/__brick__/{{name}}/pubspec.yaml
@@ -14,17 +14,12 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   intl: ^0.17.0
-
   cupertino_icons: ^1.0.5
   flutter_bloc: ^8.1.1
   equatable: ^2.0.3
   fluttertoast: ^8.0.9
   cached_network_image: ^3.2.0
-  http:
-  http_interceptor: ^1.0.2
   shared_preferences: ^2.0.12
-  retrofit: ^3.0.1+1
-  dio: ^4.0.6
   built_value: ^8.4.1
   json_annotation: ^4.6.0
   bloc: ^8.1.0
