+++ bricks/flutter_starter/__brick__/{{name}}/test/bloc/update_theme/update_theme_event_test.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: prefer_const_constructors
 
-import 'package:{{name}}/{{name}}.dart';
+import 'package:{{name}}/common_export.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
